refactor(visualization): extract visualization tools from server.py (#156)
Extract 6 visualization MCP tool functions (~724 lines) from server.py
into src/math_mcp/tools/visualization.py using the FastMCP mount pattern.

Changes:
- Create tools/visualization.py with visualization_mcp sub-server
- Add make_annotations() helper replacing inline annotation dicts (#144)
- Co-locate requires_matplotlib decorator with visualization tools
- Move evaluate_with_timeout to eval.py (shared by server and tools)
- Update tools/__init__.py to export visualization_mcp
- Mount visualization_mcp in server.py
- Update test imports to new module paths

server.py reduced from 1412 to 688 lines. All 134 tests pass.

Closes #156
Closes #144
