+++ src/math_mcp/server.py
@@ -5,15 +5,12 @@ Educational MCP server demonstrating all three MCP pillars: Tools, Resources, an
 Uses FastMCP 2.0 patterns with structured output and multi-transport support.
 """
 
-import asyncio
 import logging
 import math
-import re
-from collections.abc import AsyncIterator, Callable
+from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
 from dataclasses import dataclass
 from datetime import datetime
-from functools import wraps
 from importlib.metadata import PackageNotFoundError
 from importlib.metadata import version as pkg_version
 from typing import Annotated, Any
@@ -28,23 +25,20 @@ from fastmcp.server.middleware.rate_limiting import (
 from pydantic import Field, SkipValidation
 from starlette.responses import JSONResponse
 
-# Import visualization functions (using absolute import for FastMCP Cloud compatibility)
-from math_mcp import visualization
 from math_mcp.agent_card import AgentCard, AgentSkill
 from math_mcp.eval import (
     _classify_expression_difficulty,
     _classify_expression_topic,
     convert_temperature,
-    safe_eval_expression,
+    evaluate_with_timeout,
     validate_variable_name,
 )
 from math_mcp.settings import (
     ALLOWED_OPERATIONS,
-    ALLOWED_TRENDS,
     MathMCPSettings,
     validated_tool,
 )
-from math_mcp.tools import matrix_mcp
+from math_mcp.tools import matrix_mcp, visualization_mcp
 
 # Try importing numpy for matrix operations
 try:
@@ -61,7 +55,6 @@ except ImportError:
 settings = MathMCPSettings()
 
 # Keep constants for backward compatibility
-EXPRESSION_TIMEOUT_SECONDS = settings.math_timeout
 RATE_LIMIT_PER_MINUTE = settings.mcp_rate_limit_per_minute
 
 # === INPUT SIZE LIMITS ===
@@ -69,61 +62,7 @@ RATE_LIMIT_PER_MINUTE = settings.mcp_rate_limit_per_minute
 MAX_EXPRESSION_LENGTH = settings.max_expression_length
 MAX_STRING_PARAM_LENGTH = settings.max_string_param_length
 MAX_ARRAY_SIZE = settings.max_array_size
-MAX_GROUPS_COUNT = settings.max_groups_count
-MAX_GROUP_SIZE = settings.max_group_size
 MAX_VARIABLE_NAME_LENGTH = settings.max_variable_name_length
-MAX_DAYS_FINANCIAL = settings.max_days_financial
-
-
-def requires_matplotlib(func: Callable) -> Callable:
-    """Decorator that ensures matplotlib is available before executing visualization tools.
-
-    Wraps visualization functions to handle ImportError gracefully by returning
-    a standardized error response. Calls _setup_matplotlib() from visualization.py
-    to centralize matplotlib configuration.
-
-    This decorator eliminates ~80 lines of duplicated import/error-handling code
-    across 6 visualization tools by consolidating the pattern into a single,
-    reusable decorator.
-
-    Returns:
-        Standard error dict on ImportError with installation instructions.
-        Otherwise, executes the wrapped function normally.
-
-    Example:
-        @requires_matplotlib
-        async def plot_function(...):
-            # matplotlib is guaranteed to be available here
-            import matplotlib.pyplot as plt
-    """
-
-    @wraps(func)
-    async def wrapper(*args: Any, **kwargs: Any) -> dict[str, Any]:
-        try:
-            # Attempt to setup matplotlib; raises ImportError if unavailable
-            visualization._setup_matplotlib()
-        except ImportError:
-            # Return standardized error response
-            return {
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "**Matplotlib not available**\n\nInstall with: `pip install math-mcp-learning-server[plotting]`\n\nOr for development: `uv sync --extra plotting`",
-                        "annotations": {
-                            "error": "missing_dependency",
-                            "install_command": "pip install math-mcp-learning-server[plotting]",
-                            "difficulty": "intermediate",
-                            "topic": "visualization",
-                        },
-                    }
-                ]
-            }
-
-        # matplotlib is available; execute the wrapped function
-        return await func(*args, **kwargs)
-
-    return wrapper
-
 
 # === APPLICATION CONTEXT ===
 
@@ -155,8 +94,9 @@ mcp = FastMCP(
     instructions="A comprehensive math server demonstrating MCP fundamentals with tools, resources, and prompts for educational purposes.",
 )
 
-# Mount matrix operations tools using FastMCP composition pattern
+# Mount sub-server tools using FastMCP composition pattern
 mcp.mount(matrix_mcp)
+mcp.mount(visualization_mcp)
 
 
 # === RATE LIMITING MIDDLEWARE ===
@@ -177,56 +117,6 @@ if RATE_LIMIT_PER_MINUTE > 0:
         SlidingWindowRateLimitingMiddleware(max_requests=RATE_LIMIT_PER_MINUTE, window_minutes=1)
     )
     logging.info(f"Rate limiting enabled: {RATE_LIMIT_PER_MINUTE} requests/minute")
-
-
-# === EXPRESSION EVALUATION WITH TIMEOUT ===
-
-
-async def evaluate_with_timeout(expression: str) -> float:
-    """
-    Safely evaluate mathematical expression with execution timeout.
-
-    Prevents denial-of-service by ensuring expression evaluation completes
-    within EXPRESSION_TIMEOUT_SECONDS. Wraps synchronous safe_eval_expression()
-    in an async executor to allow timeout enforcement.
-
-    This is an educational example of wrapping CPU-bound synchronous operations
-    in async context using asyncio.wait_for() and loop.run_in_executor().
-
-    Args:
-        expression: Mathematical expression string to evaluate.
-
-    Returns:
-        float: Result of the expression evaluation.
-
-    Raises:
-        ValueError: If expression evaluation exceeds timeout or is invalid.
-    """
-    loop = asyncio.get_running_loop()
-    try:
-        return await asyncio.wait_for(
-            loop.run_in_executor(None, safe_eval_expression, expression),
-            timeout=EXPRESSION_TIMEOUT_SECONDS,
-        )
-    except TimeoutError as e:
-        raise ValueError(
-            f"Expression evaluation exceeded {EXPRESSION_TIMEOUT_SECONDS}s timeout. "
-            f"Try simplifying the expression or breaking it into smaller parts."
-        ) from e
-
-
-# === CUSTOM VALIDATORS (for logic Field constraints can't handle) ===
-
-
-def validate_nested_array_groups(groups: list[list[float]]) -> list[list[float]]:
-    """Validate nested array group sizes."""
-    for i, group in enumerate(groups):
-        if len(group) > settings.max_group_size:
-            raise ValueError(
-                f"Group {i} exceeds maximum size of {settings.max_group_size} elements. "
-                f"Current size: {len(group)}"
-            )
-    return groups
 
 
 # === TOOLS: COMPUTATIONAL OPERATIONS ===
@@ -612,620 +502,6 @@ async def load_variable(name: str, ctx: Context) -> dict[str, Any]:
             }
         ]
     }
-
-
-@mcp.tool(annotations={"title": "Function Plotter", "readOnlyHint": False, "openWorldHint": False})
-@validated_tool
-@requires_matplotlib
-async def plot_function(
-    expression: Annotated[str, Field(max_length=MAX_EXPRESSION_LENGTH)],
-    x_range: tuple[float, float],
-    num_points: Annotated[int, Field(ge=2, le=MAX_ARRAY_SIZE)] = 100,
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Generate mathematical function plots (requires matplotlib).
-
-    Args:
-        expression: Mathematical expression to plot (e.g., "x**2", "sin(x)")
-        x_range: Tuple of (min, max) for x-axis range
-        num_points: Number of points to plot (default: 100)
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        plot_function("x**2", (-5, 5))
-        plot_function("sin(x)", (-3.14, 3.14))
-    """
-
-    # FastMCP 2.0 Context logging
-    if ctx:
-        await ctx.info(f"Plotting function: {expression} over range {x_range}")
-
-    try:
-        # Validate x_range
-        x_min, x_max = x_range
-        if x_min >= x_max:
-            raise ValueError("x_range minimum must be less than maximum")
-        if num_points < 2:
-            raise ValueError("num_points must be at least 2")
-
-        # Generate x values and evaluate expression
-        import numpy as np
-
-        x_values = np.linspace(x_min, x_max, num_points)
-
-        # Evaluate expression for each x value
-        y_values = []
-        var_pattern = re.compile(r"\bx\b")
-        for x in x_values:
-            # Replace x in expression with actual value using word boundaries
-            # to avoid corrupting function names like exp, max, hex
-            expr_with_value = var_pattern.sub(f"({x})", expression)
-            try:
-                y = await evaluate_with_timeout(expr_with_value)
-                y_values.append(y)
-            except ValueError:
-                # Handle domain errors (like sqrt of negative) or timeout
-                y_values.append(float("nan"))
-
-        # Delegate rendering to visualization helper
-        image_base64 = visualization.create_function_plot(
-            x_values.tolist(), y_values, expression
-        ).decode("utf-8")
-
-        # Classify difficulty
-        difficulty = _classify_expression_difficulty(expression)
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": difficulty,
-                        "topic": "visualization",
-                        "expression": expression,
-                        "x_range": f"[{x_min}, {x_max}]",
-                        "num_points": num_points,
-                        "educational_note": "Function plotting visualizes mathematical relationships",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Plot Error:** {str(e)}\n\nPlease check your expression and x_range values.",
-                    "annotations": {
-                        "error": "plot_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-
-
-@mcp.tool(
-    annotations={"title": "Statistical Histogram", "readOnlyHint": False, "openWorldHint": False}
-)
-@validated_tool
-@requires_matplotlib
-async def create_histogram(
-    data: Annotated[list[float], Field(max_length=MAX_ARRAY_SIZE)],
-    bins: int = 20,
-    title: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Data Distribution",
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Create statistical histograms (requires matplotlib).
-
-    Args:
-        data: List of numerical values
-        bins: Number of histogram bins (default: 20)
-        title: Chart title
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        create_histogram([1, 2, 2, 3, 3, 3, 4, 4, 5], bins=5)
-    """
-    if bins < 1:
-        raise ValueError("bins must be at least 1")
-
-    # FastMCP 2.0 Context logging
-    if ctx:
-        await ctx.info(f"Creating histogram with {len(data)} data points and {bins} bins")
-
-    try:
-        # Validate inputs
-        if not data:
-            raise ValueError("Cannot create histogram with empty data")
-        if len(data) == 1:
-            raise ValueError("Histogram requires at least 2 data points")
-        if bins < 1:
-            raise ValueError("bins must be at least 1")
-
-        # Calculate statistics
-        import statistics as stats
-
-        mean_val = stats.mean(data)
-        median_val = stats.median(data)
-        std_dev = stats.stdev(data) if len(data) > 1 else 0
-
-        # Delegate rendering to visualization helper
-        image_base64 = visualization.create_histogram_chart(
-            data, bins, title, mean_val, median_val, std_dev
-        ).decode("utf-8")
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": "intermediate",
-                        "topic": "statistics",
-                        "data_points": len(data),
-                        "bins": bins,
-                        "mean": round(mean_val, 4),
-                        "median": round(median_val, 4),
-                        "std_dev": round(std_dev, 4),
-                        "educational_note": "Histograms show the distribution and frequency of data values",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Histogram Error:** {str(e)}\n\nPlease check your data and parameters.",
-                    "annotations": {
-                        "error": "histogram_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-
-
-@mcp.tool(annotations={"title": "Line Chart", "readOnlyHint": False, "openWorldHint": False})
-@validated_tool
-@requires_matplotlib
-async def plot_line_chart(
-    x_data: Annotated[list[float], Field(max_length=MAX_ARRAY_SIZE)],
-    y_data: Annotated[list[float], Field(max_length=MAX_ARRAY_SIZE)],
-    title: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Line Chart",
-    x_label: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "X",
-    y_label: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Y",
-    color: Annotated[str | None, Field(max_length=MAX_STRING_PARAM_LENGTH)] = None,
-    show_grid: bool = True,
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Create a line chart from data points (requires matplotlib).
-
-    Args:
-        x_data: X-axis data points
-        y_data: Y-axis data points
-        title: Chart title
-        x_label: X-axis label
-        y_label: Y-axis label
-        color: Line color (name or hex code, e.g., 'blue', '#2E86AB')
-        show_grid: Whether to show grid lines
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        plot_line_chart([1, 2, 3, 4], [1, 4, 9, 16], title="Squares")
-        plot_line_chart([0, 1, 2], [0, 1, 4], color='red', x_label='Time', y_label='Distance')
-    """
-
-    # Matplotlib is guaranteed to be available (decorator handles ImportError)
-    if ctx:
-        await ctx.info(f"Creating line chart with {len(x_data)} data points")
-
-    try:
-        image_base64 = visualization.create_line_chart(
-            x_data=x_data,
-            y_data=y_data,
-            title=title,
-            x_label=x_label,
-            y_label=y_label,
-            color=color,
-            show_grid=show_grid,
-        ).decode("utf-8")
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                        "chart_type": "line",
-                        "data_points": len(x_data),
-                        "educational_note": "Line charts show trends and relationships between continuous data points",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Line Chart Error:** {str(e)}\n\nPlease check that x_data and y_data have the same length and contain at least 2 points.",
-                    "annotations": {
-                        "error": "line_chart_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-
-
-@mcp.tool(annotations={"title": "Scatter Plot", "readOnlyHint": False, "openWorldHint": False})
-@validated_tool
-@requires_matplotlib
-async def plot_scatter_chart(
-    x_data: Annotated[list[float], Field(max_length=MAX_ARRAY_SIZE)],
-    y_data: Annotated[list[float], Field(max_length=MAX_ARRAY_SIZE)],
-    title: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Scatter Plot",
-    x_label: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "X",
-    y_label: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Y",
-    color: Annotated[str | None, Field(max_length=MAX_STRING_PARAM_LENGTH)] = None,
-    point_size: int = 50,
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Create a scatter plot from data points (requires matplotlib).
-
-    Args:
-        x_data: X-axis data points
-        y_data: Y-axis data points
-        title: Chart title
-        x_label: X-axis label
-        y_label: Y-axis label
-        color: Point color (name or hex code, e.g., 'blue', '#2E86AB')
-        point_size: Size of scatter points (default: 50)
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        plot_scatter_chart([1, 2, 3, 4], [1, 4, 9, 16], title="Correlation Study")
-        plot_scatter_chart([1, 2, 3], [2, 4, 5], color='purple', point_size=100)
-    """
-
-    # Matplotlib is guaranteed to be available (decorator handles ImportError)
-    if ctx:
-        await ctx.info(f"Creating scatter plot with {len(x_data)} data points")
-
-    try:
-        image_base64 = visualization.create_scatter_plot(
-            x_data=x_data,
-            y_data=y_data,
-            title=title,
-            x_label=x_label,
-            y_label=y_label,
-            color=color,
-            point_size=point_size,
-        ).decode("utf-8")
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                        "chart_type": "scatter",
-                        "data_points": len(x_data),
-                        "educational_note": "Scatter plots reveal correlations and patterns in paired data",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Scatter Plot Error:** {str(e)}\n\nPlease check that x_data and y_data have the same length.",
-                    "annotations": {
-                        "error": "scatter_plot_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "intermediate",
-                        "topic": "visualization",
-                    },
-                }
-            ]
-        }
-
-
-@mcp.tool(annotations={"title": "Box Plot", "readOnlyHint": False, "openWorldHint": False})
-@validated_tool
-@requires_matplotlib
-async def plot_box_plot(
-    data_groups: Annotated[list[list[float]], Field(max_length=MAX_GROUPS_COUNT)],
-    group_labels: Annotated[list[str] | None, Field(max_length=MAX_GROUPS_COUNT)] = None,
-    title: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Box Plot",
-    y_label: Annotated[str, Field(max_length=MAX_STRING_PARAM_LENGTH)] = "Values",
-    color: Annotated[str | None, Field(max_length=MAX_STRING_PARAM_LENGTH)] = None,
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Create a box plot for comparing distributions (requires matplotlib).
-
-    Args:
-        data_groups: List of data groups to compare
-        group_labels: Labels for each group (optional)
-        title: Chart title
-        y_label: Y-axis label
-        color: Box plot color (name or hex code, e.g., 'blue', '#2E86AB')
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        plot_box_plot([[1, 2, 3, 4], [2, 3, 4, 5]], group_labels=['Group A', 'Group B'])
-        plot_box_plot([[10, 20, 30], [15, 25, 35], [20, 30, 40]], color='green')
-    """
-    # Validate nested array group sizes
-    validate_nested_array_groups(data_groups)
-
-    # Matplotlib is guaranteed to be available (decorator handles ImportError)
-    if ctx:
-        await ctx.info(f"Creating box plot with {len(data_groups)} groups")
-
-    try:
-        image_base64 = visualization.create_box_plot(
-            data_groups=data_groups,
-            group_labels=group_labels,
-            title=title,
-            y_label=y_label,
-            color=color,
-        ).decode("utf-8")
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": "advanced",
-                        "topic": "statistics",
-                        "chart_type": "box_plot",
-                        "groups": len(data_groups),
-                        "educational_note": "Box plots display distribution quartiles, median, and outliers for comparison",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Box Plot Error:** {str(e)}\n\nPlease check that data_groups is not empty and all groups contain at least one value.",
-                    "annotations": {
-                        "error": "box_plot_error",
-                        "difficulty": "advanced",
-                        "topic": "statistics",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "advanced",
-                        "topic": "statistics",
-                    },
-                }
-            ]
-        }
-
-
-@mcp.tool(
-    annotations={"title": "Financial Line Chart", "readOnlyHint": False, "openWorldHint": False}
-)
-@validated_tool
-@requires_matplotlib
-async def plot_financial_line(
-    days: Annotated[int, Field(ge=2, le=MAX_DAYS_FINANCIAL)] = 30,
-    trend: str = "bullish",
-    start_price: float = 100.0,
-    color: Annotated[str | None, Field(max_length=MAX_STRING_PARAM_LENGTH)] = None,
-    ctx: SkipValidation[Context | None] = None,
-) -> dict[str, Any]:
-    """Generate and plot synthetic financial price data (requires matplotlib).
-
-    Creates realistic price movement patterns for educational purposes.
-    Does not use real market data.
-
-    Args:
-        days: Number of days to generate (default: 30)
-        trend: Market trend ('bullish', 'bearish', or 'volatile')
-        start_price: Starting price value (default: 100.0)
-        color: Line color (name or hex code, e.g., 'blue', '#2E86AB')
-        ctx: FastMCP context for logging
-
-    Returns:
-        Dict with base64-encoded PNG image or error message
-
-    Examples:
-        plot_financial_line(days=60, trend='bullish')
-        plot_financial_line(days=90, trend='volatile', start_price=150.0, color='orange')
-    """
-    # Validate trend against whitelist
-    if trend not in ALLOWED_TRENDS:
-        raise ValueError(f"Invalid trend: {trend}. Allowed: {', '.join(sorted(ALLOWED_TRENDS))}")
-
-    # Matplotlib is guaranteed to be available (decorator handles ImportError)
-    if ctx:
-        await ctx.info(f"Generating synthetic {trend} price data for {days} days")
-
-    try:
-        # Validate trend parameter
-        if trend not in ["bullish", "bearish", "volatile"]:
-            raise ValueError("trend must be 'bullish', 'bearish', or 'volatile'")
-
-        # Generate synthetic data
-        dates, prices = visualization.generate_synthetic_price_data(
-            days=days,
-            trend=trend,  # type: ignore
-            start_price=start_price,
-        )
-
-        # Create financial chart
-        image_base64 = visualization.create_financial_line_chart(
-            dates=dates,
-            prices=prices,
-            title=f"Synthetic {trend.capitalize()} Price Movement ({days} days)",
-            y_label="Price ($)",
-            color=color,
-        ).decode("utf-8")
-
-        # Calculate statistics
-        import statistics as stats
-
-        price_change = ((prices[-1] - prices[0]) / prices[0]) * 100
-        volatility = stats.stdev(prices) if len(prices) > 1 else 0
-
-        return {
-            "content": [
-                {
-                    "type": "image",
-                    "data": image_base64,
-                    "mimeType": "image/png",
-                    "annotations": {
-                        "difficulty": "advanced",
-                        "topic": "financial_analysis",
-                        "chart_type": "financial_line",
-                        "days": days,
-                        "trend": trend,
-                        "start_price": round(start_price, 2),
-                        "end_price": round(prices[-1], 2),
-                        "price_change_percent": round(price_change, 2),
-                        "volatility": round(volatility, 2),
-                        "educational_note": "Synthetic data generated for educational purposes only - not real market data",
-                    },
-                }
-            ]
-        }
-
-    except ValueError as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Financial Chart Error:** {str(e)}\n\nPlease check your parameters (days >= 2, valid trend, positive start_price).",
-                    "annotations": {
-                        "error": "financial_chart_error",
-                        "difficulty": "advanced",
-                        "topic": "financial_analysis",
-                    },
-                }
-            ]
-        }
-    except Exception as e:
-        return {
-            "content": [
-                {
-                    "type": "text",
-                    "text": f"**Unexpected Error:** {str(e)}",
-                    "annotations": {
-                        "error": "unexpected_error",
-                        "difficulty": "advanced",
-                        "topic": "financial_analysis",
-                    },
-                }
-            ]
-        }
 
 
 # === RESOURCES: DATA EXPOSURE ===
